refactor(*): move builtin amp components to builtins/ folder
