+++ builtins/amp-img.md
